% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-formatter.R
\name{block_structure_gemini}
\alias{block_structure_gemini}
\title{Build Gemini-Style Blocks}
\usage{
block_structure_gemini(prompts_df, api, system = NULL, ...)
}
\description{
Build Gemini-Style Blocks
}
\keyword{internal}
