% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude-batch.R
\name{claude_check_batch_status}
\alias{claude_check_batch_status}
\title{Retrieve the status of a specific Message Batch Using the Anthropic API}
\usage{
claude_check_batch_status(
  batch_response,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{batch_response}{The list returned by \code{\link[=claude_create_batch]{claude_create_batch()}}, containing the string
representing the unique ID of the message batch to retrieve. (It's also
possible to pass the specific object from \code{\link[=claude_list_batches]{claude_list_batches()}}, e.g.,
\code{claude_list_batches()$data[[1]]}).}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing information about the specified message batch,
including processing status and result URLs.
}
\description{
This function retrieves details for a specific message batch using the
\href{https://docs.anthropic.com/en/api/message-batches-beta}{Message Batch endpoint}. It can be
used to poll for batch completion or to access the batch metadata.
}
