% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude-batch.R
\name{claude_list_batches}
\alias{claude_list_batches}
\title{List Message Batches Using the Anthropic API}
\usage{
claude_list_batches(
  limit = 20,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{limit}{An integer specifying the maximum number of batches to return.
Defaults to 20, and ranges from 1 to 100.}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing information about any message batches,
including batch IDs, processing status, and URLs for retrieving results.
}
\description{
This function retrieves all message batches within a workspace that are
currently in process. It uses the \href{https://docs.anthropic.com/en/api/message-batches-beta}{List Message Batches endpoint}. The most
recently created batches are returned first.
}
