% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude-batch.R
\name{claude_poll_and_retrieve_results}
\alias{claude_poll_and_retrieve_results}
\title{Poll Message Batch Status and Retrieve Results}
\usage{
claude_poll_and_retrieve_results(
  batch_response,
  timeout = 3600,
  max_retries = 3,
  pause_cap = 1200,
  tidy = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{batch_response}{A list returned by \code{\link[=claude_create_batch]{claude_create_batch()}} or
\code{\link[=claude_check_batch_status]{claude_check_batch_status()}} containing batch details. (It's also
possible to pass the specific object from \code{\link[=claude_list_batches]{claude_list_batches()}}, e.g.,
\code{claude_list_batches()$data[[1]]}).}

\item{timeout}{A numeric value representing the time (in seconds) to wait
between polling attempts. Defaults to 3600 seconds (1 hour).}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{tidy}{A logical value indicating whether to attempt to tidy the
resulting json into a tidy \link{tibble}. Default is \code{TRUE}; \code{FALSE} is useful
if you want to do the tidying separately or prefer the raw json.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A character vector containing each line of the .jsonl result file if
the batch ends successfully.
}
\description{
This function continuously polls the status of a message batch until the
processing is complete. If the status is "ended", it downloads the results.
If the status is "canceling", it stops and provides an informative message.
If the status is "in_progress", it waits for the specified timeout before
polling again.
}
