% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude.R
\name{claude_single_request}
\alias{claude_single_request}
\title{Send a Single Request to the Anthropic API}
\usage{
claude_single_request(
  prompt,
  model,
  n_candidates = 1,
  max_retries = 10,
  temperature = 0.2,
  max_tokens = 300,
  json_mode = FALSE,
  system = NULL,
  response_validation_fun,
  content_extraction_fun,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{prompt}{A list containing the prompt messages, created using a specific
format understood by the Anthropic API, e.g., \code{list(list(role = "user", content = "hello world"))}.}

\item{model}{The model to be used for generating the response, such as
"claude-3-haiku".}

\item{n_candidates}{The number of response candidates to generate. Defaults
to 1.}

\item{max_retries}{An integer specifying the maximum number of retry attempts
in case of request failures. Defaults to 10.}

\item{temperature}{A numeric value controlling the randomness of the model's
output. A higher value makes the output more random. Defaults to 0.2.}

\item{max_tokens}{The maximum number of tokens to generate for the response.
Defaults to 300.}

\item{json_mode}{A logical value indicating whether to process any json
generated by the response. Defaults to \code{FALSE}.}

\item{system}{A character string providing additional context or instructions
for the system. Defaults to \code{NULL}.}

\item{response_validation_fun}{A function to validate the response content.
If not provided, the default validation function is used.}

\item{content_extraction_fun}{A function to extract the content from the
response. If not provided, a default extraction function is used based on
\code{json_mode}.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A tibble containing the response content and usage details, including
prompt and response tokens.
}
\description{
This function sends a single prompt to the Anthropic API for message
completion. It handles retries in case of request failures and returns the
response content.
}
