% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude-token-count.R
\name{claude_token_count}
\alias{claude_token_count}
\title{Count Tokens for a Given Message Using Anthropic API}
\usage{
claude_token_count(
  messages,
  model,
  system = NULL,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{messages}{A list of prompts created by \code{\link[=build_prompts_from_files]{build_prompts_from_files()}}.}

\item{model}{A string specifying the model to be used (e.g.,
"claude-3-5-sonnet-20241022").}

\item{system}{(Optional) A string specifying a system prompt to provide
context and instructions to the model. Tokens in the system prompt are
counted as well.}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
An integer representing the total number of input tokens across the
provided list of messages and system prompt.
}
\description{
This function calls the Anthropic Token Count API to determine the number of
tokens in a given message. The API is useful for understanding token usage in
models before sending the actual message for processing.
}
\examples{
messages <- list(
  list(
    list(
      role = "user", content = "Hello, world"
      )
    )
   )
class(messages) <- c("claude", class(messages))
model <- "claude-3-5-sonnet-20241022"
tokens <- claude_token_count(messages, model)
print(tokens)
}
