% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_readers.R
\name{crew_log_retries}
\alias{crew_log_retries}
\title{Extract Log Data for API Retries}
\usage{
crew_log_retries(path = ".", glob = "crew-*.log")
}
\arguments{
\item{path}{The path to the directory containing log files. Defaults to the
current directory (".").}

\item{glob}{A pattern to match the log files. Defaults to "crew-*.log".}
}
\value{
A tibble containing log data with columns: \code{file}, \code{error_codes}, and
\code{retry_interval}. If no logs are found, returns an empty tibble.
}
\description{
This function extracts log data related to failed API requests created by the
package \link{crew} when running in a \link{targets} pipeline from the specified
directory. It looks for lines indicating failed requests and retrieves the
error codes and retry intervals.
}
\examples{
# Extract retry log data from the current directory
crew_log_retries()

}
