% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt-formatter.R
\name{format_character}
\alias{format_character}
\title{Format a character vector for dispatch to embedding, tokenize, completion,
and detokenize APIs.}
\usage{
format_character(api, content)
}
\arguments{
\item{api}{The specific api to use, such as "openai".}

\item{content}{A character vector.}
}
\description{
Format a character vector for dispatch to embedding, tokenize, completion,
and detokenize APIs.
}
