% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utilities.R
\name{get_available_models}
\alias{get_available_models}
\title{Get Available Models for an API}
\usage{
get_available_models(api, mode = "chat")
}
\arguments{
\item{api}{An optional character string specifying the API for which the
available models are needed. Must be one of "groq", "claude", "openai",
"mistral", or "gemini". If not provided, the function returns models for
all APIs.}

\item{mode}{The model use: must be one of \code{"chat"}, \code{"embedding"}, or
\code{"rerank"}.}
}
\value{
A character vector containing the available models for the specified
API, or all available models if no API is specified.
}
\description{
This function returns the available models for a specified API from the
internal \code{models_df} data frame that contains supported models and their
associated APIs. If no API is specified, it returns all available models
across all APIs.
}
\details{
The function checks the \code{models_df} data frame to find the models
associated with the specified API. If no API is provided, it returns all
models from \code{models_df}. Typing \code{R.AI:::models_df} should produce the
models dataframe.
}
\examples{
# Get available models for the OpenAI API
get_available_models("openai")

# Get all available models across all APIs
get_available_models()

}
\seealso{
\code{\link[=get_default_model]{get_default_model()}} for retrieving the default model for a
specified API.

Other model utilities: 
\code{\link{get_available_apis}()},
\code{\link{get_default_model}()},
\code{\link{is_api_key_available}()},
\code{\link{list_models}()}
}
\concept{model utilities}
