% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utilities.R
\name{get_default_model}
\alias{get_default_model}
\title{Get the Default Model for an API}
\usage{
get_default_model(api, type = "cheapest")
}
\arguments{
\item{api}{A character string specifying the API for which the default model
is needed. Must be one of "groq", "claude", "openai", "mistral", or
"gemini".}

\item{type}{A character string specifying the type of model to return. Must
be one of "cheapest", "largest", or "best". Default is "cheapest".}
}
\value{
A character string representing the default model for the specified
API and type.
}
\description{
This function returns the default model for a specified API based on the
provided type. It selects the model from the internal \code{models_df} data frame
that contains supported models and their associated APIs.
}
\details{
The function checks the \code{models_df} data frame to find the models
associated with the specified API. It then selects a model based on the
given type:
\itemize{
\item "cheapest": Selects the model with the lowest input cost.
\item "largest": Selects the largest model based on predefined choices for
each API.
\item "best": Selects the best model based on predefined choices for each API.
}
}
\examples{
# Get the cheapest model for the OpenAI API
get_default_model("openai", type = "cheapest")

# Get the largest model for the GROQ API
get_default_model("groq", type = "largest")

}
\seealso{
\code{\link[=get_available_models]{get_available_models()}} for retrieving all available models for an
API and \code{\link[=get_available_apis]{get_available_apis()}} for retrieving all available APIs.

Other model utilities: 
\code{\link{get_available_apis}()},
\code{\link{get_available_models}()},
\code{\link{is_api_key_available}()},
\code{\link{list_models}()}
}
\concept{model utilities}
