% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamafile.R
\name{is_llamafile_running}
\alias{is_llamafile_running}
\title{Check if Llamafile is Running on localhost:8080}
\usage{
is_llamafile_running()
}
\value{
TRUE if any Llamafile instance is running, otherwise FALSE.
}
\description{
This function checks if a Llamafile instance is currently running at
localhost:8080.
}
