% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral-batch.R
\name{mistral_cancel_batch_job}
\alias{mistral_cancel_batch_job}
\title{Cancel Mistral Batch Job}
\usage{
mistral_cancel_batch_job(
  job_id,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{job_id}{The batch job ID to cancel.}

\item{max_retries}{An integer indicating the maximum number of retry attempts in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the response details after the cancel request is submitted.
}
\description{
This function cancels an ongoing batch request using the Mistral Batch API.
}
