% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral-batch.R
\name{mistral_check_batch_status}
\alias{mistral_check_batch_status}
\title{Check Status of Mistral Batch Job}
\usage{
mistral_check_batch_status(
  job_id,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{job_id}{The batch job ID to check the status for.}

\item{max_retries}{An integer indicating the maximum number of retry attempts in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the current status of the batch.
}
\description{
This function checks the status of an existing batch request using the Mistral Batch API.
}
