% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral-batch.R
\name{mistral_download_batch_results}
\alias{mistral_download_batch_results}
\title{Download Results of Mistral Batch Job}
\usage{
mistral_download_batch_results(
  output_file_id,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{output_file_id}{The output file ID to download the results for.}

\item{max_retries}{An integer indicating the maximum number of retry attempts in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress messages during retries. Defaults to \code{FALSE}.}
}
\value{
A character vector containing each line of the .jsonl result file.
}
\description{
This function downloads the results of a completed batch request using the output file ID from the Mistral Batch API.
}
