% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistral-batch.R
\name{mistral_upload_batch_file}
\alias{mistral_upload_batch_file}
\title{Upload JSONL File to Mistral for Batch Processing}
\usage{
mistral_upload_batch_file(
  jsonl,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{jsonl}{The JSONL string generated by \code{\link[=mistral_format_prompts_for_batch]{mistral_format_prompts_for_batch()}}.}

\item{max_retries}{An integer indicating the maximum number of retry attempts in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the response details, including the file ID.
}
\description{
This function uploads a JSONL file to Mistral for batch processing using the Files API.
}
