% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-threads.R
\name{openai_create_thread}
\alias{openai_create_thread}
\alias{openai_get_thread}
\alias{openai_modify_thread}
\alias{openai_delete_thread}
\title{Manage Threads in OpenAI}
\usage{
openai_create_thread(messages = NULL, tool_resources = NULL, metadata = NULL)

openai_get_thread(thread_id)

openai_modify_thread(thread_id, tool_resources = NULL, metadata = NULL)

openai_delete_thread(thread_id)
}
\arguments{
\item{messages}{List. Optional. A list of messages to start the thread with.
Each message should include properties such as \code{role} and \code{content}. These
can be generated with \code{\link[=prompt_list]{prompt_list()}}.}

\item{tool_resources}{List or \code{NULL}. Optional. A set of resources made
available to the assistant's tools in this thread.}

\item{metadata}{List. Optional. A set of 16 key-value pairs to attach to the
thread. Keys can be a maximum of 64 characters long, and values can be a
maximum of 512 characters long.}

\item{thread_id}{Character. Required for thread retrieval, modification, or
deletion. The ID of the thread to interact with.}
}
\value{
Depending on the function used:
\itemize{
\item \code{openai_create_thread}: A list containing the created thread object.
\item \code{openai_get_thread}: The details of the specified thread.
\item \code{openai_modify_thread}: The updated thread object with modified metadata or tool resources.
\item \code{openai_delete_thread}: The deletion status, including whether the thread was successfully deleted.
}
}
\description{
These functions interact with the OpenAI \href{https://platform.openai.com/docs/api-reference/threads}{Threads API}. See the
documentation there for more details.
}
\seealso{
Other openai: 
\code{\link{openai_batch_job}()},
\code{\link{openai_cancel_batch}()},
\code{\link{openai_check_batch_status}()},
\code{\link{openai_create_assistant}()},
\code{\link{openai_create_message}()},
\code{\link{openai_create_run}()},
\code{\link{openai_create_vector_store}()},
\code{\link{openai_create_vector_store_file}()},
\code{\link{openai_create_vector_store_file_batch}()},
\code{\link{openai_download_batch_results}()},
\code{\link{openai_list_batches}()},
\code{\link{openai_list_run_steps}()},
\code{\link{openai_poll_and_download}()},
\code{\link{openai_submit_tool_outputs}()},
\code{\link{openai_upload_file}()}

Other threads: 
\code{\link{openai_create_message}()}
}
\concept{openai}
\concept{threads}
