% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-batch.R
\name{openai_check_batch_status}
\alias{openai_check_batch_status}
\title{Check Status of OpenAI Batch Request}
\usage{
openai_check_batch_status(
  batch_response,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{batch_response}{The response object returned by \code{\link[=openai_create_batch]{openai_create_batch()}}
containing the batch ID.}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the current status of the batch, including the
status field.
}
\description{
This function checks the status of an existing batch request using the OpenAI
Batch API.
}
