% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-batch.R
\name{openai_create_batch}
\alias{openai_create_batch}
\title{Create OpenAI Batch Request}
\usage{
openai_create_batch(
  upload_response,
  endpoint = "/v1/chat/completions",
  completion_window = "24h",
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{upload_response}{The response object returned by
\code{\link[=openai_upload_batch_file]{openai_upload_batch_file()}} containing the uploaded file ID.}

\item{endpoint}{The endpoint for processing each request in the batch, e.g.,
"/v1/chat/completions".}

\item{completion_window}{A string specifying the completion window, currently
must be "24h".}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the response details, including batch ID and
status.
}
\description{
This function creates a batch request using the uploaded file ID from the
OpenAI Batch API.
}
