% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-batch.R
\name{openai_download_batch_results}
\alias{openai_download_batch_results}
\title{Download Results of OpenAI Batch Request}
\usage{
openai_download_batch_results(
  batch_status,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE,
  tidy = TRUE
)
}
\arguments{
\item{batch_status}{The status object returned by
\code{\link[=openai_check_batch_status]{openai_check_batch_status()}} or \code{\link[=openai_create_batch]{openai_create_batch()}} containing the
\code{output_file_id} field.}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}

\item{tidy}{A logical value indicating whether to attempt to tidy the
resulting json into a tidy \link{tibble}. Default is \code{TRUE}; \code{FALSE} is useful
if you want to do the tidying separately or prefer the raw json.}
}
\value{
A character vector containing each line of the .jsonl result file.
}
\description{
This function downloads the results of a completed batch request using the
output file ID from the OpenAI Batch API.
}
