% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-batch.R
\name{openai_list_batches}
\alias{openai_list_batches}
\title{List OpenAI Batches}
\usage{
openai_list_batches(
  limit = 10,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{limit}{An integer specifying the maximum number of batches to return.
Defaults to 10.}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the details of the batches.
}
\description{
This function lists all the batches being processed for a user using the
OpenAI Batch API.
}
\seealso{
Other openai: 
\code{\link{openai_batch_job}()},
\code{\link{openai_cancel_batch}()},
\code{\link{openai_check_batch_status}()},
\code{\link{openai_create_assistant}()},
\code{\link{openai_create_message}()},
\code{\link{openai_create_run}()},
\code{\link{openai_create_thread}()},
\code{\link{openai_create_vector_store}()},
\code{\link{openai_create_vector_store_file}()},
\code{\link{openai_create_vector_store_file_batch}()},
\code{\link{openai_download_batch_results}()},
\code{\link{openai_list_run_steps}()},
\code{\link{openai_poll_and_download}()},
\code{\link{openai_submit_tool_outputs}()},
\code{\link{openai_upload_file}()}

Other batch: 
\code{\link{batch_job}()},
\code{\link{call_batch_api}()},
\code{\link{cancel_batch}()},
\code{\link{check_batch_status}()},
\code{\link{claude_batch_job}()},
\code{\link{claude_cancel_batch}()},
\code{\link{claude_check_batch_status}()},
\code{\link{claude_download_batch_results}()},
\code{\link{claude_list_batches}()},
\code{\link{claude_poll_and_download}()},
\code{\link{download_results}()},
\code{\link{list_batches}()},
\code{\link{openai_batch_job}()},
\code{\link{openai_cancel_batch}()},
\code{\link{openai_check_batch_status}()},
\code{\link{openai_download_batch_results}()},
\code{\link{openai_poll_and_download}()},
\code{\link{poll_and_download}()}
}
\concept{batch}
\concept{openai}
