% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-batch.R
\name{openai_poll_and_retrieve_results}
\alias{openai_poll_and_retrieve_results}
\title{Poll OpenAI Batch Status and Retrieve Results}
\usage{
openai_poll_and_retrieve_results(
  batch_response,
  timeout = 3600,
  max_retries = 3,
  pause_cap = 1200,
  tidy = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{batch_response}{A list returned by \code{openai_create_batch} or
\code{openai_check_batch_status} containing batch details.}

\item{timeout}{A numeric value representing the time (in seconds) to wait
between polling attempts. Defaults to 3600 seconds (1 hour).}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{tidy}{A logical value indicating whether to attempt to tidy the
resulting JSON into a tidy \link{tibble}. Default is \code{TRUE}; \code{FALSE} is useful
if you want to do the tidying separately or prefer the raw JSON.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A character vector containing each line of the .jsonl result file if
the batch completes successfully.
}
\description{
This function continuously polls the status of a batch until processing is
complete. If the status is "completed" or "expired", it downloads the
results. If the status is "in_progress", "validating", or "finalizing", it
waits for the specified timeout before polling again. If the status is
"cancelling", "cancelled", or "failed, it stops and provides an informative
message.
}
