% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openai-run.R
\name{openai_submit_tool_outputs}
\alias{openai_submit_tool_outputs}
\title{Submit Tool Outputs to a Run in OpenAI}
\usage{
openai_submit_tool_outputs(thread_id, run_id, tool_outputs, stream = NULL)
}
\arguments{
\item{thread_id}{Character. Required. The ID of the thread to which the run
belongs.}

\item{run_id}{Character. Required. The ID of the run that requires tool
output submission.}

\item{tool_outputs}{List. Required. A list of tool outputs to submit. Each
item should include \code{tool_call_id} and \code{output}.}

\item{stream}{Logical or NULL. Optional. If TRUE, streams events during the
submission process.}
}
\value{
A list containing the updated run object.
}
\description{
Submit the outputs from tool calls for a specific run when the run requires
action.
}
\seealso{
Other run: 
\code{\link{openai_create_run}()},
\code{\link{openai_list_run_steps}()}

Other openai: 
\code{\link{openai_batch_job}()},
\code{\link{openai_cancel_batch}()},
\code{\link{openai_check_batch_status}()},
\code{\link{openai_create_assistant}()},
\code{\link{openai_create_message}()},
\code{\link{openai_create_run}()},
\code{\link{openai_create_thread}()},
\code{\link{openai_create_vector_store}()},
\code{\link{openai_create_vector_store_file}()},
\code{\link{openai_create_vector_store_file_batch}()},
\code{\link{openai_download_batch_results}()},
\code{\link{openai_list_batches}()},
\code{\link{openai_list_run_steps}()},
\code{\link{openai_poll_and_download}()},
\code{\link{openai_upload_file}()}

Other assistants: 
\code{\link{openai_create_assistant}()},
\code{\link{openai_list_run_steps}()}
}
\concept{assistants}
\concept{openai}
\concept{run}
