% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_builders.R
\name{prompt}
\alias{prompt}
\title{Create a Single Prompt for API Requests}
\usage{
prompt(user, api, system, data)
}
\arguments{
\item{user}{A character string containing the user's input prompt.}

\item{api}{A string specifying the API to use. Must be one of \code{"gemini"},
\code{"openai"}, \code{"claude"}, \code{"groq"}, \code{"llamafile"}, or \code{"mistral"}.}

\item{system}{Optional. A character string providing system-level
instructions or context for the API.}

\item{data}{Optional. A data frame containing variables to be interpolated
into the \code{user} content using \code{stringr::str_glue_data()}.}
}
\value{
A structured prompt for use with single-request API functions like
\code{\link[=openai_single_request]{openai_single_request()}} or \code{\link[=claude_single_request]{claude_single_request()}}. The exact structure
depends on the specified API.
}
\description{
This function creates a single prompt for use with single-request API
functions. It processes the input \code{user} content and combines it with
optional \code{system} instructions, returning a structured prompt compatible with
the specified API.
}
\details{
This function processes a single user's input prompt along with
optional system instructions, ensuring compatibility with the target API.
It is designed for single-request functions like \code{openai_single_request},
\code{gemini_single_request}, and \code{claude_single_request}.
}
\examples{
# Create a single prompt for the OpenAI API
single_prompt <- prompt(
  user = "What are the benefits of exercise?",
  api = "openai",
  system = "You are a helpful assistant."
)
print(single_prompt)

}
\seealso{
\itemize{
\item \code{\link[=prompt_list]{prompt_list()}} for creating prompt lists for use with \code{\link[=call_api]{call_api()}}.
\item \code{\link[=call_api]{call_api()}} for executing API calls with generated prompts.
}

Other prompts: 
\code{\link{build_prompts_from_files}()},
\code{\link{prompt_list}()}
}
\concept{prompts}
