% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.AI-package.R
\docType{package}
\name{R.AI-package}
\alias{R.AI}
\alias{R.AI-package}
\title{R.AI: Use LLM APIs from within R}
\description{
This package contains functions to call a number of LLM APIs, including OpenAI, Groq, Anthropic (Claude), and Gemini. It also helps manage your prompts.
}
\author{
\strong{Maintainer}: Xavier Marquez \email{marquez.x@gmail.com} (\href{https://orcid.org/0000-0001-7653-9334}{ORCID})

}
\keyword{internal}
