% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voyage.R
\name{voyage_reranker}
\alias{voyage_reranker}
\title{Voyage AI Reranker}
\usage{
voyage_reranker(
  query,
  documents,
  model,
  top_k = NULL,
  return_documents = FALSE,
  truncation = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{query}{\code{character} A string containing the query. The query must not be
empty.}

\item{documents}{\code{character} A character vector containing the documents to
be reranked. Must contain at least one document and no more than 1000.}

\item{model}{\code{character} The name of the reranking model to use. Supported
options are:
\itemize{
\item \code{"rerank-2"} - Allows up to 4000 query tokens, 16000 tokens per document, and 300,000 total tokens.
\item \code{"rerank-2-lite"} - Allows up to 2000 query tokens, 8000 tokens per document, and 300,000 total tokens.
\item \code{"rerank-1"} - Allows up to 2000 query tokens, 8000 tokens per document, and 300,000 total tokens.
\item \code{"rerank-lite-1"} - Allows up to 1000 query tokens, 4000 tokens per document, and 300,000 total tokens.
}}

\item{top_k}{\code{integer}, optional The number of most relevant documents to
return. If not specified, reranking results for all documents will be
returned.}

\item{return_documents}{\code{logical}, optional Whether to include the documents
in the response. Defaults to \code{FALSE}.
\itemize{
\item If \code{FALSE}, the API returns a list of indices and relevance scores.
\item If \code{TRUE}, the API also includes the documents in the response.
}}

\item{truncation}{\code{logical}, optional Whether to truncate the input to
satisfy the token limits. Defaults to \code{TRUE}.
\itemize{
\item If \code{TRUE}, the query and documents will be truncated as needed to meet the limits.
\item If \code{FALSE}, an error will be raised if the token limits are exceeded.
}}

\item{quiet}{\code{logical}, optional If \code{TRUE}, suppresses logging messages.
Defaults to \code{FALSE}.}
}
\value{
A \code{tibble} containing the reranking results with the following
columns:
\itemize{
\item \code{index}: The index of the document in the input list.
\item \code{relevance_score}: The relevance score of the document with respect to the query.
\item \code{document}: (optional) The document string, included if \code{return_documents = TRUE}.
}
}
\description{
Sends a query and a list of documents to the Voyage AI reranking API and
retrieves the reranking results.
}
\details{
This function communicates with the Voyage AI reranking API to
prioritize documents based on their relevance to the input query.

Logging can be controlled using the \code{quiet} parameter. API key must be set in
the \code{VOYAGE_API_KEY} environment variable.
}
\seealso{
\code{\link[=call_embedding_api]{call_embedding_api()}} for embedding queries and documents.
}
