% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamafile.R
\name{which_llamafile_running}
\alias{which_llamafile_running}
\title{Determines which llamfile model is running}
\usage{
which_llamafile_running()
}
\value{
The name of the runnning llamafile model, or \code{NA} if no llamafile is
running.
}
\description{
Determines which llamafile model is running, if any.
}
\seealso{
Other llamafile: 
\code{\link{is_llamafile_running}()},
\code{\link{kill_llamafile}()},
\code{\link{llamafile_single_request}()},
\code{\link{start_llamafile}()}
}
\concept{llamafile}
