% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero-helpers.R
\name{zotero_collection_attachments}
\alias{zotero_collection_attachments}
\title{Get PDF Attachments from a Zotero Collection}
\usage{
zotero_collection_attachments(collection_name, zotero_dir)
}
\arguments{
\item{collection_name}{A character string specifying the name of the Zotero
collection from which to retrieve PDF attachments.}

\item{zotero_dir}{The directory where your Zotero database lives. If not
specified, the function will try to guess.}
}
\value{
A \code{\link[=tibble]{tibble()}} (data frame) with an added class "zotero_collection"
containing:
\itemize{
\item \code{libraryID}: The library ID associated with the item.
\item \code{collectionID}: The collection ID from Zotero.
\item \code{itemID}: The ID of the item in Zotero.
\item \code{collectionKey}: The collection key, useful to retrieve it from the zotero API.
\item \code{attachmentItemID}: The attachment ID.
\item \code{attachmentKey}: The subfolder (attachment key) within the Zotero \code{storage} directory.
\item \code{collectionName}: The collection name in Zotero.
\item \code{path}: The full file path to the PDF attachment on the local system.
\item \code{itemKey}: The key of the item in Zotero - useful for the API.
\item \code{fileLastModified}: When the file was last modified.
And, if \code{better-bibtex} was installed:
\item \code{citationKey}: The \code{better-bibtex} citationKey for the item.
\item \code{pinned}: Whether the citation key is pinned or auto-generated.
}
}
\description{
This function retrieves all PDF attachments in a specified Zotero collection.
It connects to the Zotero database, identifies items within the given
collection, and then locates associated PDF files stored in Zotero's
\code{storage} directory. Finally, it returns a \code{\link[=tibble]{tibble()}} including full file
paths. The zotero database must be closed before using this function;
otherwise it will throw an error. The function also requires the \code{DBI} and
\code{RSQLite} packages to be installed.
}
\examples{
\dontrun{
# Retrieve PDF attachments from a collection named "My Research"
pdfs <- zotero_collection_attachments("My Research")
head(pdfs)
}
}
\concept{zotero helpers}
