% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-embeddings.R
\name{call_embedding_api}
\alias{call_embedding_api}
\title{Call Embedding API}
\usage{
call_embedding_api(texts, model, api, ...)
}
\arguments{
\item{texts}{\code{character} A character vector containing the input texts to
process. Must not be empty.}

\item{model}{\code{character} A string specifying the embedding model to use. The
model must be supported by the chosen API.}

\item{api}{\code{character} Specifies the API or method for embedding generation.
Supported values are:
\itemize{
\item \code{"mistral"} - Dispatches to the Mistral embedding API.
\item \code{"openai"} - Dispatches to the OpenAI embedding API.
\item \code{"cohere"} - Dispatches to the Cohere embedding API.
\item \code{"llamafile"} - Dispatches to a local LlamaFile instance.
}}

\item{...}{Additional arguments passed to API-specific implementations. See
details for supported options.}
}
\value{
A \code{tibble} containing the embeddings and metadata, with the following
columns:
\itemize{
\item \code{id}: Identifier for each input text.
\item \code{text_set}: Hash representing the batch of texts processed together.
\item \code{embedding}: A list-column containing the embedding vector for each input text.
}
}
\description{
A generic function to retrieve text embeddings by sending requests to various
APIs or a local LlamaFile instance.
}
\details{
The function dispatches requests to API-specific methods based on
the value of \code{api}.

Additional arguments for specific APIs include:
\itemize{
\item \code{llamafile}: \code{llamafile_path} (path to the LlamaFile model), \code{quiet} (logical, suppresses logging).
\item \code{cohere}: \code{input_type} (type of input for embeddings, e.g., "search_document"), \code{embedding_types} (list of desired embedding types, e.g., "float"), \code{quiet} (logical, suppresses logging).
\item \code{openai} and \code{mistral}: No additional arguments beyond \code{quiet}.
}
}
\seealso{
\code{\link[=list_models]{list_models()}} for retrieving available models for specific APIs.
}
