% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude-batch.R
\name{claude_batch_job}
\alias{claude_batch_job}
\title{Create a Message Batch Using the Anthropic API}
\usage{
claude_batch_job(prompts, model, max_tokens = 300, quiet = FALSE)
}
\arguments{
\item{prompts}{A list of prompts created by \code{\link[=build_prompts_from_files]{build_prompts_from_files()}}.}

\item{model}{The model to use to process the batch. Currently supports only
Claude 3 Haiku, Claude 3 Opus, and Claude 3.5 Sonnet. Defaults to
"claude-3-haiku-20240307" (i.e., \code{get_default_model("claude", type = "cheapest")}.}

\item{max_tokens}{The maximum number of output tokens for each item in the
batch.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\description{
This function sends a batch of message creation requests to the Anthropic API
using the \href{https://docs.anthropic.com/en/api/creating-message-batches}{Message Batches endpoint}. It
allows you to process multiple requests at once, useful for scenarios
requiring bulk message completions.
}
