% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude-batch.R
\name{claude_cancel_batch}
\alias{claude_cancel_batch}
\title{Cancel a Message Batch Using the Anthropic API}
\usage{
claude_cancel_batch(
  batch_response,
  max_retries = 3,
  pause_cap = 1200,
  quiet = FALSE
)
}
\arguments{
\item{batch_response}{The list returned by \code{\link[=claude_create_batch]{claude_create_batch()}}, containing the string
representing the unique ID of the message batch to retrieve. (It's also
possible to pass the specific object from \code{\link[=claude_list_batches]{claude_list_batches()}}, e.g.,
\code{claude_list_batches()$data[[1]]}).}

\item{max_retries}{An integer indicating the maximum number of retry attempts
in case of request failures. Defaults to 3.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A list containing the response details, including batch ID,
processing status, and counts of canceled requests.
}
\description{
This function cancels a message batch that is currently in progress using the
\href{https://docs.anthropic.com/en/api/message-batches-beta}{Cancel a Message Batch endpoint}. The batch
may enter a canceling state, during which any in-progress, non-interruptible
requests may still complete.
}
