% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claude.R
\name{claude_single_request}
\alias{claude_single_request}
\title{Send a Single Request to the Anthropic API}
\usage{
claude_single_request(
  prompt,
  model,
  max_retries = 10,
  temperature = 0.2,
  max_tokens = 300,
  json_mode = FALSE,
  system = NULL,
  content_extraction_fun,
  pause_cap = 1200,
  quiet = FALSE,
  stop_sequences = NULL
)
}
\arguments{
\item{prompt}{A list containing the prompt message(s). Typically this is the
output of \code{\link[=prompt]{prompt()}}.}

\item{model}{A string specifying the model to use. Get available models with
\code{\link[=list_models]{list_models()}} or \code{\link[=get_available_models]{get_available_models()}}.}

\item{max_retries}{The maximum number of retry attempts in case of request
failures. Defaults to 10.}

\item{temperature}{A numeric value between 0 and 1 that controls the
randomness of the response. Higher values make the output more random.
Defaults to 0.2.}

\item{max_tokens}{The maximum number of tokens to include in the response.
Defaults to 300.}

\item{json_mode}{A logical value indicating whether the response should be
parsed as JSON. Defaults to \code{FALSE}.}

\item{system}{A character string providing additional context or instructions
for the system. If not provided, the function will attempt to extract it
from the attribute \code{system} in \code{prompt} (if it exists); otherwise it will
default to \code{NULL}.}

\item{content_extraction_fun}{A function to extract the desired content from
the API response. If not provided, a default extraction function is used
depending on the value of \code{json_mode}.}

\item{pause_cap}{A numeric value representing the maximum pause duration (in
seconds) between retries. Defaults to 1200.}

\item{quiet}{A logical value indicating whether the function should suppress
messages during retries. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[=tibble]{tibble()}} with a \code{response} column and usage information.
}
\description{
This function sends a single prompt to the \href{https://docs.anthropic.com/en/api/messages}{Anthropic API} for messages.It also handles
retries and extracts the relevant response content.
}
\details{
See the Anthropic API documentation at
\url{https://docs.anthropic.com/en/api/messages}
for more.
}
\seealso{
Other claude: 
\code{\link{claude_batch_job}()},
\code{\link{claude_cancel_batch}()},
\code{\link{claude_check_batch_status}()},
\code{\link{claude_download_batch_results}()},
\code{\link{claude_list_batches}()},
\code{\link{claude_poll_and_download}()}

Other single message: 
\code{\link{groq_single_request}()},
\code{\link{llamafile_single_request}()},
\code{\link{mistral_single_request}()},
\code{\link{openai_single_request}()}
}
\concept{claude}
\concept{single message}
