% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_readers.R
\name{crew_log_summary}
\alias{crew_log_summary}
\title{Summarize Log Data}
\usage{
crew_log_summary(path = ".", glob = "crew-*.log")
}
\arguments{
\item{path}{The path to the directory containing log files. Defaults to the
current directory (".").}

\item{glob}{A pattern to match the log files. Defaults to "crew-*.log".}
}
\value{
A tibble summarizing log data with columns: \code{model}, \code{api},
\code{prompt_name}, \code{prompt_set}, \code{file}, \code{prompts_processed}, \code{first_activity},
\code{last_activity}, \code{total_time}, \code{pph}, \code{eta}, \code{last_error_code}, and
\code{last_retry_interval}.
}
\description{
This function provides a summary of log data created by the package \link{crew}
when running in a \link{targets} pipeline extracted from the specified directory.
It includes information such as the number of prompts processed, the time of
the first and last activity, the total processing time, and an estimate of
prompts processed per hour (pph).
}
\examples{
# Summarize log data from the current directory
crew_log_summary()

}
