% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamafile.R
\name{kill_llamafile}
\alias{kill_llamafile}
\title{Kill the Running Llamafile Process}
\usage{
kill_llamafile()
}
\value{
TRUE if the Llamafile process was successfully terminated, FALSE otherwise.
}
\description{
This function checks if a Llamafile instance is running and attempts to terminate it.
}
\seealso{
Other llamafile: 
\code{\link{is_llamafile_running}()},
\code{\link{start_llamafile}()},
\code{\link{which_llamafile_running}()}
}
\concept{llamafile}
