% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama-cpp-api.R
\name{health.llama_cpp}
\alias{health.llama_cpp}
\alias{props.llama_cpp}
\title{Functions to interact with the llama.cpp server API}
\usage{
health.llama_cpp(...)

props.llama_cpp(...)
}
\arguments{
\item{...}{Additional arguments passed to the HTTP request.}
}
\description{
See
\url{https://github.com/ggerganov/llama.cpp/tree/master/examples/server}
for more details.
}
