% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llamafile.R
\name{start_llamafile}
\alias{start_llamafile}
\title{Start Llamafile Instance}
\usage{
start_llamafile(
  llamafile_path = here::here("models", "Llama-3.2-1B-Instruct.Q6_K.llamafile.exe")
)
}
\arguments{
\item{llamafile_path}{The path to the Llamafile executable.}
}
\description{
This function attempts to start the Llamafile executable if it is not already running.
}
